edited show.html.erb

+++ lib/tasks/dev.rake
@@ -1,14 +1,26 @@
 desc "Fill the database tables with some sample data"
 task({ :sample_data => :environment }) do
   puts "Sample data task running"
-  
-  ActiveRecord::Base.connection.tables.each do |t|
-    ActiveRecord::Base.connection.reset_pk_sequence!(t)
+  if Rail.env.developement?
+    User.destroy_all
+    Board.destroy_all
+    Post.destroy_all
   end
+  if Rails.env.production?
+    ActiveRecord::Base.connection.tables.each do |t|
+      ActiveRecord::Base.connection.reset_pk_sequence!(t)
+    end
+  end
+    #new user creation
+  usernames = ["alice", "bob", "eve", "carol", "dave"]
 
-  Board.destroy_all
-  Post.destroy_all
-  
+  usernames.each do |username|
+  user = User.new
+  user.email = "#{username}@example.com"
+  user.password = "password"
+  user.save
+end
+
   5.times do
     board = Board.new
     board.name = Faker::Address.community
